separated courseEmitter from eventMatcher
