+++ src/events/emitter.js
@@ -1,0 +1,5 @@
+const EventEmitter = require('events');
+
+class Emitter extends EventEmitter { }
+
+module.exports = Emitter;
