+++ src/events/eventMatcher.js
@@ -1,13 +1,4 @@
-const EventEmitter = require('events');
-const request = require('request-promise-native');
-const config = require('config');
-
-const { courseModel } = require('../services/user-group/model');
-
-class Emitter extends EventEmitter { }
-const courseEmitter = new Emitter();
-
-const REQUEST_TIMEOUT = 8000;
+const courseEmitter = require('./courseEmitter');
 
 const EventMatcher = {
 
@@ -42,68 +33,5 @@ const EventMatcher = {
     }
 
 };
-
-function sendCourseUpdatedPushMessage(course, receivers) {
-
-    if(!receivers) return Promise.reject('receivers missing!');
-
-    // todo retrieve config from app context instead
-    const serviceUrls = config.get('services') || {};
-    const notification = config.get('notification') || {};
-
-    const data = {
-        "notification": {
-            "title": `Kurs ${course.name} wurde aktualisiert.`,
-            "body": "Neue Inhalte werden automatisch geladen."
-        },
-        "data": {
-            "tag": "course-data-updated",
-            "courseId": course._id,
-        },
-        "receivers": receivers,
-        "template": "tpl",
-        "languagePayloads": "lp"
-    };
-
-    const options = {
-        uri: serviceUrls.notification + '/push/',
-        method: 'POST',
-        headers: {
-            //'token': userId
-        },
-        body: Object.assign({}, data,
-            { serviceUrl: serviceUrls.notification },
-            { platformId: notification.platformId }
-        ),
-        json: true,
-        timeout: REQUEST_TIMEOUT
-    };
-
-    return request(options).then(response => {
-        return response;
-    });
-}
-
-courseEmitter.on('updated', (courseId) => {
-    console.log('course updated event...', courseId);
-
-    // todo get relevant ids of course
-    return courseModel.findById(courseId).then(course => {
-        if (!course) return Promise.reject('courseId not found');
-        let userIds = new Set();
-        course.teacherIds.forEach(id => userIds.add(id));
-        course.userIds.forEach(id => userIds.add(id));
-
-        // todo send post message
-        return sendCourseUpdatedPushMessage(course, [...userIds]);
-    }).catch(err => console.log(err));
-
-});
-
-courseEmitter.on('removed', (course) => {
-
-});
-
-
 
 module.exports = EventMatcher;
